added assert service program
The assert service program was extracted from the ILEUnit project (originally
from the RPGUnit project) and setup as a standalone module/service program.
It depends on the MESSAGE service program which has already been added to
the OSSILE project. An example on how to use the assert service program has
also been included.