Initial support for STRTCPSVR *GUNICORN handler
Adds SERVER(*GUNICORN) to STRTCPSVR/ENDTCPSVR. This allows an easy deployment mechanism for Python applications. It includes support for INSTANTCE(*AUTOSTART) to have them started at STRTCP time as well.

An ini file can be created under /QOpenSys/etc/gunicorn-tcpsvr/INSTANCE.ini, where INSTANCE is the name specified on the INSTANCE() parameter to STRTCPSVR. Various options exist. See the comment at the top of gunicorn_tcp.c for how to create one. Note that this file must be created in UTF-8.


See setup for how to register with ADDTCPSVR or copy the command which is printed during the build.